refactor: core 디렉토리 생성
