+++ src/main/java/lotto/core/BonusNumber.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.core;
 
 import lotto.common.PureNumber;
