+++ src/main/java/lotto/core/DrawResult.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.core;
 
 public enum DrawResult {
