+++ src/main/java/lotto/core/LotteryTicket.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.core;
 
 import java.util.ArrayList;
 import java.util.List;
