+++ src/main/java/lotto/core/PortfolioReport.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.core;
 
 public class PortfolioReport {
