+++ src/test/java/lotto/BonusNumberTest.java
@@ -3,6 +3,8 @@ package lotto;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 import java.util.List;
+import lotto.core.BonusNumber;
+import lotto.core.LotteryTicket;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
