+++ src/test/java/lotto/LotteryTicketTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.core.LotteryTicket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
